+++ Logic/joint.py
@@ -1,0 +1,7 @@
+class Joint:
+	"""
+	Joint class
+	"""
+	def __init__(self, x, y):
+		self.posx = x
+		self.posy = y
